Relocated ItineraryManager
